+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 #### Changes
-- [BREAKING] `Process` no longer takes ownership of the `Host` (#1563)
+- [BREAKING] `Process` no longer takes ownership of the `Host` (#1571)
+- [BREAKING] `ProcessState` was converted from a trait to a struct (#1571)
 
 ## 0.11.0 (2024-11-04)